refactor: change the top level switch/case in reqHandler
