+++ .prettierrc.json
@@ -1,7 +1,7 @@
 {
   "endOfLine": "auto",
   "semi": true,
-  "singleQuote": true,
+  "singleQuote": false,
   "tabWidth": 2,
   "trailingComma": "es5",
   "printWidth": 100
